Comecei menu em modo texto.

+++ Loja/src/loja/Estoque.java
@@ -35,7 +35,7 @@ public class Estoque {
         return -1;
     }
     
-    public boolean AlteraEstoque(int codigo,int Novocodigo,int quant,int preco_uni,String Descri,String Categ,String particu){  //retorna 1 caso alteracoes sejam realizadas, 0 caso não.
+    public boolean AlteraEstoque(int codigo,int Novocodigo,int quant,double preco_uni,String Descri,String Categ,String particu){  //retorna 1 caso alteracoes sejam realizadas, 0 caso não.
         int a = ProcuraIndex(codigo);
         if(a>-1){
             return Estoque.get(a).SetProduto(Novocodigo, quant, preco_uni, Descri, Categ, particu);
